File restructure of GW models
