+++ src/gateway_api/endpoints/transaction_endpoints.rs
@@ -4,6 +4,10 @@ impl GatewayClient {
     /// Get Construction Metadata
     ///
     /// Returns information needed to construct a new transaction including current epoch number.
+    ///
+    /// See [the Gateway API docs for details][doc].
+    ///
+    /// [doc]: https://radix-babylon-gateway-api.redoc.ly/#operation/TransactionConstruction
     pub(crate) async fn transaction_construction(&self) -> Result<LedgerState> {
         self.post_empty(
             "transaction/construction",
@@ -14,18 +18,30 @@ impl GatewayClient {
         .await
     }
 
-    /// Preview Transaction
+    /// Previews transaction against the network - aka "dry run" of transaction.
     ///
-    /// Previews transaction against the network. This endpoint is effectively a
-    /// proxy towards the Core API /v0/transaction/preview endpoint.
+    /// Previews are used for [`execution_summary`][TransactionManifest::execution_summary], which the iOS/Android wallet app
+    /// uses to present the "review transaction" screen to the user.
     ///
-    /// See [the Core API documentation][core] for more details.
+    /// See [the Gateway API docs for details][doc].
     ///
-    /// [core]: https://radix-babylon-core-api.redoc.ly/#tag/Transaction/paths/~1transaction~1preview/post
+    /// [doc]: https://radix-babylon-gateway-api.redoc.ly/#operation/TransactionPreview
     pub(crate) async fn transaction_preview(
         &self,
         request: TransactionPreviewRequest,
     ) -> Result<TransactionPreviewResponse> {
         self.post("transaction/preview", request, res_id).await
+    }
+
+    /// Submits a signed transaction payload to the network.
+    ///
+    /// See [the Gateway API docs for details][doc].
+    ///
+    /// [doc]: https://radix-babylon-gateway-api.redoc.ly/#operation/TransactionSubmit
+    pub(crate) async fn transaction_submit(
+        &self,
+        request: TransactionSubmitRequest,
+    ) -> Result<TransactionSubmitResponse> {
+        self.post("transaction/submit", request, res_id).await
     }
 }
