+++ src/gateway_api/methods/transaction_methods.rs
@@ -13,7 +13,7 @@ impl GatewayClient {
     /// transaction - a preview of the transaction. The "EncodedReceipt" is
     /// required by the [`execution_summary` method](TransactionManifest::execution_summary)
     /// on [`TransactionManifest`].
-    pub async fn transaction_dry_run(
+    pub async fn dry_run_transaction(
         &self,
         intent: TransactionIntent,
         signer_public_keys: Vec<PublicKey>,
@@ -24,5 +24,23 @@ impl GatewayClient {
             .await
             .map(|r| r.encoded_receipt)
             .and_then(|s| BagOfBytes::from_str(&s))
+    }
+
+    pub async fn submit_notarized_transaction(
+        &self,
+        notarized_transaction: NotarizedTransaction,
+    ) -> Result<bool> {
+        // let compiled = notarized_transaction.compile();
+        // let request = TransactionSubmitRequest::new(compiled);
+        // self.transaction_submit()
+        todo!()
+    }
+}
+
+impl FungibleResourcesCollectionItem {
+    pub fn resource_address(&self) -> ResourceAddress {
+        match self {
+            Self::Global(item) => item.resource_address,
+        }
     }
 }
