+++ src/gateway_api/models/mod.rs
@@ -1,5 +1,2 @@
-mod request;
-mod response;
-
-pub use request::*;
-pub use response::*;
+mod types;
+pub use types::*;
