--- src/gateway_api/models/request/mod.rs
@@ -1,5 +1,0 @@
-mod state_entity_details;
-mod transaction_preview;
-
-pub use state_entity_details::*;
-pub use transaction_preview::*;
