--- src/gateway_api/models/request/state_entity_details.rs
@@ -1,6 +1,0 @@
-use crate::prelude::*;
-
-#[derive(Clone, Debug, PartialEq, Eq, Serialize)]
-pub struct StateEntityDetailsRequest {
-    pub(crate) addresses: Vec<Address>,
-}
