--- src/gateway_api/models/response/transaction_construction.rs
@@ -1,6 +1,0 @@
-use crate::prelude::*;
-
-#[derive(Deserialize, Clone, PartialEq, Eq, Debug, uniffi::Record)]
-pub struct TransactionConstructionResponse {
-    pub ledger_state: LedgerState,
-}
