+++ src/gateway_api/models/types/mod.rs
@@ -1,0 +1,5 @@
+mod request;
+mod response;
+
+pub use request::*;
+pub use response::*;
