+++ src/gateway_api/models/types/request/state/entity/state_entity_details.rs
@@ -1,0 +1,13 @@
+use crate::prelude::*;
+
+#[derive(
+    Clone,
+    Debug,
+    PartialEq,
+    Eq,
+    Serialize,
+    Deserialize, /* Deserialize so we can test roundtrip of JSON vectors */
+)]
+pub struct StateEntityDetailsRequest {
+    pub(crate) addresses: Vec<Address>,
+}
