+++ src/gateway_api/models/types/request/transaction/preview/mod.rs
@@ -1,0 +1,3 @@
+mod transaction_preview;
+
+pub use transaction_preview::*;
