+++ src/gateway_api/models/types/request/transaction/submit/transaction_submit.rs
@@ -1,0 +1,14 @@
+use crate::prelude::*;
+
+#[derive(
+    Clone,
+    Debug,
+    PartialEq,
+    Eq,
+    Serialize,
+    Deserialize, /* Deserialize so we can test roundtrip of JSON vectors */
+)]
+pub struct TransactionSubmitRequest {
+    /** Hex-encoded notarized transaction payload which can be submitted. */
+    notarized_transaction_hex: String,
+}
