+++ src/gateway_api/models/types/response/state/entity/details/fungible/collection.rs
@@ -1,0 +1,15 @@
+use crate::prelude::*;
+
+#[derive(
+    Deserialize,
+    Serialize, /* Serialize so we can test roundtrip of JSON vectors */
+    Clone,
+    PartialEq,
+    Eq,
+    Debug,
+    uniffi::Record,
+)]
+pub struct FungibleResourcesCollection {
+    pub items: Vec<FungibleResourcesCollectionItem>,
+}
+
