+++ src/gateway_api/models/types/response/state/entity/details/fungible/collection_item.rs
@@ -1,0 +1,17 @@
+use crate::prelude::*;
+
+#[derive(
+    Deserialize,
+    Serialize, /* Serialize so we can test roundtrip of JSON vectors */
+    Clone,
+    PartialEq,
+    Eq,
+    Debug,
+    EnumAsInner,
+    uniffi::Enum,
+)]
+#[serde(untagged)]
+pub enum FungibleResourcesCollectionItem {
+    Global(FungibleResourcesCollectionItemGloballyAggregated),
+}
+
