+++ src/gateway_api/models/types/response/state/entity/details/fungible/collection_item_global.rs
@@ -1,0 +1,16 @@
+use crate::prelude::*;
+
+#[derive(
+    Deserialize,
+    Serialize,
+    /* Serialize so we can test roundtrip of JSON vectors */
+    Clone,
+    PartialEq,
+    Eq,
+    Debug,
+    uniffi::Record,
+)]
+pub struct FungibleResourcesCollectionItemGloballyAggregated {
+    pub amount: Decimal192,
+    pub resource_address: ResourceAddress,
+}
