+++ src/gateway_api/models/types/response/state/entity/details/mod.rs
@@ -1,0 +1,7 @@
+mod fungible;
+mod state_entity_details_response;
+mod state_entity_details_response_item;
+
+pub use fungible::*;
+pub use state_entity_details_response::*;
+pub use state_entity_details_response_item::*;
