+++ src/gateway_api/models/types/response/state/entity/details/state_entity_details_response_item.rs
@@ -1,0 +1,15 @@
+use crate::prelude::*;
+
+#[derive(
+    Deserialize,
+    Serialize, /* Serialize so we can test roundtrip of JSON vectors */
+    Clone,
+    PartialEq,
+    Eq,
+    Debug,
+    uniffi::Record,
+)]
+pub struct StateEntityDetailsResponseItem {
+    pub address: Address,
+    pub fungible_resources: Option<FungibleResourcesCollection>,
+}
