+++ src/gateway_api/models/types/response/state/entity/mod.rs
@@ -1,0 +1,2 @@
+mod details;
+pub use details::*;
