+++ src/gateway_api/models/types/response/state/mod.rs
@@ -1,0 +1,3 @@
+mod entity;
+
+pub use entity::*;
