+++ src/gateway_api/models/types/response/transaction/construction/mod.rs
@@ -1,0 +1,3 @@
+mod transaction_construction_response;
+
+pub use transaction_construction_response::*;
