+++ src/gateway_api/models/types/response/transaction/construction/transaction_construction_response.rs
@@ -1,0 +1,14 @@
+use crate::prelude::*;
+
+#[derive(
+    Deserialize,
+    Serialize, /* Serialize so we can test roundtrip of JSON vectors */
+    Clone,
+    PartialEq,
+    Eq,
+    Debug,
+    uniffi::Record,
+)]
+pub struct TransactionConstructionResponse {
+    pub ledger_state: LedgerState,
+}
