+++ src/gateway_api/models/types/response/transaction/preview/logs_inner.rs
@@ -1,14 +1,8 @@
 use crate::prelude::*;
 
-#[derive(Deserialize, Clone, PartialEq, Eq, Debug, uniffi::Record)]
-pub struct TransactionPreviewResponse {
-    /** Hex-encoded binary blob. */
-    pub encoded_receipt: String,
-    pub logs: Vec<TransactionPreviewResponseLogsInner>,
-}
-
 #[derive(
     Deserialize,
+    Serialize, /* Serialize so we can test roundtrip of JSON vectors */
     Clone,
     PartialEq,
     Eq,
