+++ src/gateway_api/models/types/response/transaction/preview/mod.rs
@@ -1,0 +1,5 @@
+mod transaction_preview_response;
+mod logs_inner;
+
+pub use transaction_preview_response::*;
+pub use logs_inner::*;
