+++ src/gateway_api/models/types/response/transaction/submit/mod.rs
@@ -1,0 +1,3 @@
+mod submit;
+
+pub use submit::*;
